7topdig/trunk/requirement.txt
FLask
Babel
Flask-DebugToolbar
pytz
Flask-WTF
lamson
Flask-SQLAlchemy
docutils
markdown
feedparser
